+++ sdk/spring/azure-spring-data-cosmos/src/test/java/com/azure/spring/data/cosmos/common/TestConstants.java
@@ -134,6 +134,10 @@ public final class TestConstants {
     public static final int PRECONDITION_FAILED_STATUS_CODE = 412;
     public static final int CONFLICT_STATUS_CODE = 409;
 
+    public static final Long LongId1 = new Long(10000001);
+    public static final Long LongId2 = new Long(20000002);
+    public static final Long LongId3 = new Long(54600003);
+
     public static final Map<String, String> PASSPORT_IDS_BY_COUNTRY = new HashMap<String, String>() {{
         put("United States of America", "123456789");
         put("Côte d'Ivoire", "IC1234567");